Move code from `HomeUpdated` to `Home` and delete `HomeUpdated`
